+++ changes.md
@@ -35,3 +35,14 @@
 - fixed image disproportion in the recommeded products in market page
 - changed the layout for maketpage (products and services)
 - The featured and most recent tabs in the services page now works
+
+-
+- Chat panel now scrolls to the bottom when you send a new message
+- You can now favorite a chat and also show only favorited chats
+- Fixed the message blurred issue that was caused because of improper git merging
+- Added ( You) to a chat and chat header when the chat is the same user
+
+-
+- Bookmarks for services and products now work
+- Improved the UI/UX of the bookmarks page
+- You can now also see the company that owns a product/service that you bookedmarked